Add fixtures to ubench.h.
This commit adds fixtures to ubench.h via a similar mechanism to how they work for utest.h. Fixtures are a great way of initializing data once and then using it in the body of the benchmark - without skewing the data collection of the benchmark with all the initialization code.